Fix clang-tidy suggestion

+++ src/fheroes2/heroes/heroes_move.cpp
@@ -532,7 +532,7 @@ std::vector<std::pair<fheroes2::Point, fheroes2::Sprite>> Heroes::getHeroSprites
 
     int flagFrameID = sprite_index;
     if ( !isMoveEnabled() ) {
-        flagFrameID = isShipMaster() ? 0 : Game::getAdventureMapAnimationIndex();
+        flagFrameID = isShipMaster() ? 0 : static_cast<int>( Game::getAdventureMapAnimationIndex() );
     }
 
     fheroes2::Point offset;
